+++ src/Resources/app/administration/src/provider/open_id_connect/decorator/condition-type-data-provider.decorator.ts
@@ -1,0 +1,12 @@
+import {HeptacomOauthRuleDataProviderService} from '../../../service/heptacom-oauth-rule-data-provider.service';
+
+export default (ruleConditionService: HeptacomOauthRuleDataProviderService) => {
+    ruleConditionService.addCondition('heptacomAdminOpenAuthAuthenticatedRequest', {
+        component: 'heptacom-admin-open-auth-condition-authenticated-request',
+        label: 'heptacomAdminOpenAuthClient.providerFields.open_id_connect.condition.rule.authenticatedRequest',
+        scopes: ['open_id_connect', 'cidaas', 'google_cloud', 'keycloak', 'microsoft_azure_oidc', 'okta', 'onelogin'],
+        group: 'user',
+    });
+
+    return ruleConditionService;
+};